+++ src/client/event.cpp
@@ -278,7 +278,7 @@ float mir_surface_output_event_get_scale(MirSurfaceOutputEvent const* ev)
     return ev->scale();
 }
 
-float mir_surface_output_event_get_refresh_rate(MirSurfaceOutputEvent const* ev)
+double mir_surface_output_event_get_refresh_rate(MirSurfaceOutputEvent const* ev)
 {
     expect_event_type(ev, mir_event_type_surface_output);
     return ev->refresh_rate();
